chore: removed unecessary trait
